png

Module adam7

Utility functions related to handling of the Adam7 algorithm.

Structs

Adam7Info

Describes which stage of the Adam7 algorithm applies to a decoded row.

Adam7Iterator 🔒

This iterator iterates over the different passes of an image Adam7 encoded PNG image The pattern is: 16462646 77777777 56565656 77777777 36463646 77777777 56565656 77777777

Functions

expand_adam7_bits 🔒

Given row_stride, interlace info, and bits-per-pixel, produce an iterator of bit positions of pixels to copy from the input scanline to the image buffer. The positions are expressed as bit offsets from position (0,0) in the frame that is currently being decoded.

expand_pass

Copies pixels from interlaced_row into the right location in img.

subbyte_pixels 🔒