png::adam7

Function expand_adam7_bits

fn expand_adam7_bits(
    row_stride_in_bytes: usize,
    info: &Adam7Info,
    bits_pp: u8,
) -> impl Iterator<Item = BitPostion>

Given row_stride, interlace info, and bits-per-pixel, produce an iterator of bit positions of pixels to copy from the input scanline to the image buffer. The positions are expressed as bit offsets from position (0,0) in the frame that is currently being decoded.

This should only be used with bits_pp < 8.