png

Module common

Common types shared between the encoder and decoder

Structs

AnimationControl

Animation control information

CodingIndependentCodePoints

Coding-independent code points (cICP) specify the color space (primaries), transfer function, matrix coefficients and scaling factor of the image using the code points specified in ITU-T-H.273.

ContentLightLevelInfo

Content light level information of HDR content.

FrameControl

Frame control information

Info

PNG info struct

MasteringDisplayColorVolume

Mastering Display Color Volume (mDCV) used at the point of content creation, as specified in SMPTE-ST-2086.

ParameterError

PixelDimensions

Pixel dimensions information

ScaledFloat

An unsigned integer scaled version of a floating point value, equivalent to an integer quotient with fixed denominator).

SourceChromaticities

Chromaticities of the color space primaries

Transformations

Output transformations

Enums

BitDepth

Bit depth of the PNG file. Specifies the number of bits per sample.

BlendOp

How pixels are written into the buffer.

BytesPerPixel 🔒

Internal count of bytes per pixel. This is used for filtering which never uses sub-byte units. This essentially reduces the number of possible byte chunk lengths to a very small set of values appropriate to be defined as an enum.

ColorType

Describes how a pixel is encoded.

Compression

The type and strength of applied compression.

DeflateCompression

Advanced compression settings with more customization options than Compression.

DisposeOp

How to reset buffer of an animated png (APNG) at the end of a frame.

ParameterErrorKind 🔒

SrgbRenderingIntent

The rendering intent for an sRGB image.

Unit

Physical unit of the pixel dimensions