Struct png::common::Info

pub struct Info {
    pub width: u32,
    pub height: u32,
    pub bit_depth: BitDepth,
    pub color_type: ColorType,
    pub interlaced: bool,
    pub trns: Option<Vec<u8>>,
    pub pixel_dims: Option<PixelDimensions>,
    pub palette: Option<Vec<u8>>,
    pub frame_control: Option<FrameControl>,
    pub animation_control: Option<AnimationControl>,
    pub compression: Compression,
    pub filter: FilterType,
}

PNG info struct

Fields
width: u32height: u32bit_depth: BitDepthcolor_type: ColorTypeinterlaced: booltrns: Option<Vec<u8>>pixel_dims: Option<PixelDimensions>palette: Option<Vec<u8>>frame_control: Option<FrameControl>animation_control: Option<AnimationControl>compression: Compressionfilter: FilterType
Implementations

impl Info

pub fn size(&self) -> (u32, u32)

Size of the image

pub fn is_animated(&self) -> bool

Returns true if the image is an APNG image.

pub fn animation_control(&self) -> Option<&AnimationControl>

Returns the frame control information of the image

pub fn frame_control(&self) -> Option<&FrameControl>

Returns the frame control information of the current frame

pub fn bits_per_pixel(&self) -> usize

Returns the bits per pixel

pub fn bytes_per_pixel(&self) -> usize

Returns the bytes per pixel

pub(crate) fn bpp_in_prediction(&self) -> BytesPerPixel

Return the number of bytes for this pixel used in prediction.

Some filters use prediction, over the raw bytes of a scanline. Where a previous pixel is require for such forms the specification instead references previous bytes. That is, for a gray pixel of bit depth 2, the pixel used in prediction is actually 4 pixels prior. This has the consequence that the number of possible values is rather small. To make this fact more obvious in the type system and the optimizer we use an explicit enum here.

pub fn raw_bytes(&self) -> usize

Returns the number of bytes needed for one deinterlaced image

pub fn raw_row_length(&self) -> usize

Returns the number of bytes needed for one deinterlaced row

pub(crate) fn checked_raw_row_length(&self) -> Option<usize>

pub fn raw_row_length_from_width(&self, width: u32) -> usize

Returns the number of bytes needed for one deinterlaced row of width width

impl Info

fn validate(&self, fc: &FrameControl) -> Result<(), DecodingError>