png::encoder

Struct ChunkWriter

struct ChunkWriter<'a, W: Write> {
    writer: ChunkOutput<'a, W>,
    buffer: Vec<u8>,
    index: usize,
    curr_chunk: ChunkType,
}

This writer is used between the actual writer and the ZlibEncoder and has the job of packaging the compressed data into a PNG chunk, based on the image metadata

Currently the way it works is that the specified buffer will hold one chunk at the time and buffer the incoming data until flush is called or the maximum chunk size is reached.

The maximum chunk is the smallest between the selected buffer size and u32::MAX >> 1 (0x7fffffff or 2147483647 dec)

When a chunk has to be flushed the length (that is now known) and the CRC will be written at the correct locations in the chunk.

Fields
writer: ChunkOutput<'a, W>buffer: Vec<u8>index: usize

keeps track of where the last byte was written

curr_chunk: ChunkType
Implementations

impl<'a, W: Write> ChunkWriter<'a, W>

fn new(writer: ChunkOutput<'a, W>, buf_len: usize) -> ChunkWriter<'a, W>

fn next_frame_info(&self) -> (usize, usize)

Returns the size of each scanline for the next frame paired with the size of the whole frame

This is used by the StreamWriter to know when the scanline ends so it can filter compress it and also to know when to start the next one

fn write_header(&mut self) -> Result<(), EncodingError>

NOTE: this bypasses the internal buffer so the flush method should be called before this in the case there is some data left in the buffer when this is called, it will panic

fn set_fctl(&mut self, f: FrameControl)

Set the FrameControl for the following frame

It will ignore the sequence_number of the parameter as it is updated internally.

fn flush_inner(&mut self) -> Result<()>

Flushes the current chunk