png

Enum Decoded

pub enum Decoded {
    Nothing,
    ChunkBegin(u32, ChunkType),
    ChunkComplete(ChunkType),
    BadAncillaryChunk(ChunkType),
    SkippedAncillaryChunk(ChunkType),
    ImageData,
    ImageDataFlushed,
}

Result of the decoding process

Variants

Nothing

Nothing decoded yet

ChunkBegin(u32, ChunkType)

A chunk header (length and type fields) has been read.

ChunkComplete(ChunkType)

Chunk has been read successfully.

BadAncillaryChunk(ChunkType)

An ancillary chunk has been read but it was in the wrong place, had corrupt contents, or had an invalid CRC.

SkippedAncillaryChunk(ChunkType)

Skipped an ancillary chunk because it was unrecognized or the decoder was configured to skip this type of chunk.

ImageData

Decoded raw image data.

ImageDataFlushed

The last of a consecutive chunk of IDAT was done. This is distinct from ChunkComplete which only marks that some IDAT chunk was completed but not that no additional IDAT chunk follows.