Struct png::Decoder

pub struct Decoder<R: Read> { /* private fields */ }

PNG Decoder

Implementations

impl<R: Read> Decoder<R>

pub fn new(r: R) -> Decoder<R>

Create a new decoder configuration with default limits.

pub fn new_with_limits(r: R, limits: Limits) -> Decoder<R>

Create a new decoder configuration with custom limits.

pub fn new_with_options(r: R, decode_options: DecodeOptions) -> Decoder<R>

Create a new decoder configuration with custom DecodeOptions.

pub fn set_limits(&mut self, limits: Limits)

Limit resource usage.

Note that your allocations, e.g. when reading into a pre-allocated buffer, are NOT considered part of the limits. Nevertheless, required intermediate buffers such as for singular lines is checked against the limit.

Note that this is a best-effort basis.

use std::fs::File;
use png::{Decoder, Limits};
// This image is 32×32, 1bit per pixel. The reader buffers one row which requires 4 bytes.
let mut limits = Limits::default();
limits.bytes = 3;
let mut decoder = Decoder::new_with_limits(File::open("tests/pngsuite/basi0g01.png").unwrap(), limits);
assert!(decoder.read_info().is_err());

// This image is 32x32 pixels, so the decoder will allocate less than 10Kib
let mut limits = Limits::default();
limits.bytes = 10*1024;
let mut decoder = Decoder::new_with_limits(File::open("tests/pngsuite/basi0g01.png").unwrap(), limits);
assert!(decoder.read_info().is_ok());

pub fn read_header_info(&mut self) -> Result<&Info<'static>, DecodingError>

Read the PNG header and return the information contained within.

Most image metadata will not be read until read_info is called, so those fields will be None or empty.

pub fn read_info(self) -> Result<Reader<R>, DecodingError>

Reads all meta data until the first IDAT chunk

pub fn set_transformations(&mut self, transform: Transformations)

Set the allowed and performed transformations.

A transformation is a pre-processing on the raw image data modifying content or encoding. Many options have an impact on memory or CPU usage during decoding.

pub fn set_ignore_text_chunk(&mut self, ignore_text_chunk: bool)

Set the decoder to ignore all text chunks while parsing.

eg.

use std::fs::File;
use png::Decoder;
let mut decoder = Decoder::new(File::open("tests/pngsuite/basi0g01.png").unwrap());
decoder.set_ignore_text_chunk(true);
assert!(decoder.read_info().is_ok());

pub fn set_ignore_iccp_chunk(&mut self, ignore_iccp_chunk: bool)

Set the decoder to ignore iccp chunks while parsing.

eg.

use std::fs::File;
use png::Decoder;
let mut decoder = Decoder::new(File::open("tests/iccp/broken_iccp.png").unwrap());
decoder.set_ignore_iccp_chunk(true);
assert!(decoder.read_info().is_ok());

pub fn ignore_checksums(&mut self, ignore_checksums: bool)

Set the decoder to ignore and not verify the Adler-32 checksum and CRC code.