Struct png::Encoder

pub struct Encoder<W: Write> { /* fields omitted */ }

PNG Encoder

Implementations

impl<W: Write> Encoder<W>

pub fn new(w: W, width: u32, height: u32) -> Encoder<W>

pub fn set_palette(&mut self, palette: Vec<u8>)

pub fn set_trns(&mut self, trns: Vec<u8>)

pub fn write_header(self) -> Result<Writer<W>, EncodingError>

pub fn set_color(&mut self, color: ColorType)

Set the color of the encoded image.

These correspond to the color types in the png IHDR data that will be written. The length of the image data that is later supplied must match the color type, otherwise an error will be emitted.

pub fn set_depth(&mut self, depth: BitDepth)

Set the indicated depth of the image data.

pub fn set_compression<C: Into<Compression>>(&mut self, compression: C)

Set compression parameters.

Accepts a Compression or any type that can transform into a Compression. Notably deflate::Compression and deflate::CompressionOptions which “just work”.

pub fn set_filter(&mut self, filter: FilterType)

Set the used filter type.

The default filter is FilterType::Sub which provides a basic prediction algorithm for sample values based on the previous. For a potentially better compression ratio, at the cost of more complex processing, try out FilterType::Paeth.