Struct png::StreamWriter

pub struct StreamWriter<'a, W: Write> { /* private fields */ }

Streaming PNG writer

This may silently fail in the destructor, so it is a good idea to call finish or flush before dropping.

Implementations

impl<'a, W: Write> StreamWriter<'a, W>

pub fn set_filter(&mut self, filter: FilterType)

Set the used filter type for the next frame.

The default filter is FilterType::Sub which provides a basic prediction algorithm for sample values based on the previous.

For optimal compression ratio you should enable adaptive filtering instead of setting a single filter for the entire image, see set_adaptive_filter.

pub fn set_adaptive_filter(&mut self, adaptive_filter: AdaptiveFilterType)

Set the adaptive filter type for the next frame.

Adaptive filtering attempts to select the best filter for each line based on heuristics which minimize the file size for compression rather than use a single filter for the entire image.

The default method is AdaptiveFilterType::NonAdaptive.

pub fn set_frame_delay( &mut self, numerator: u16, denominator: u16, ) -> Result<(), EncodingError>

Set the fraction of time the following frames are going to be displayed, in seconds

If the denominator is 0, it is to be treated as if it were 100 (that is, the numerator then specifies 1/100ths of a second). If the value of the numerator is 0 the decoder should render the next frame as quickly as possible, though viewers may impose a reasonable lower bound.

This method will return an error if the image is not animated.

pub fn set_frame_dimension( &mut self, width: u32, height: u32, ) -> Result<(), EncodingError>

Set the dimension of the following frames.

This function will return an error when:

The image is not an animated;

The selected dimension, considering also the current frame position, goes outside the image boundaries;

One or both the width and height are 0;

pub fn set_frame_position( &mut self, x: u32, y: u32, ) -> Result<(), EncodingError>

Set the position of the following frames.

An error will be returned if:

The image is not animated;

The selected position, considering also the current frame dimension, goes outside the image boundaries;

pub fn reset_frame_dimension(&mut self) -> Result<(), EncodingError>

Set the frame dimension to occupy all the image, starting from the current position.

To reset the frame to the full image size reset_frame_position should be called first.

This method will return an error if the image is not animated.

pub fn reset_frame_position(&mut self) -> Result<(), EncodingError>

Set the frame position to (0, 0).

Equivalent to calling set_frame_position(0, 0).

This method will return an error if the image is not animated.

pub fn set_blend_op(&mut self, op: BlendOp) -> Result<(), EncodingError>

Set the blend operation for the following frames.

The blend operation specifies whether the frame is to be alpha blended into the current output buffer content, or whether it should completely replace its region in the output buffer.

See the BlendOp documentation for the possible values and their effects.

Note that for the first frame the two blend modes are functionally equivalent due to the clearing of the output buffer at the beginning of each play.

This method will return an error if the image is not animated.

pub fn set_dispose_op(&mut self, op: DisposeOp) -> Result<(), EncodingError>

Set the dispose operation for the following frames.

The dispose operation specifies how the output buffer should be changed at the end of the delay (before rendering the next frame)

See the DisposeOp documentation for the possible values and their effects.

Note that if the first frame uses DisposeOp::Previous it will be treated as DisposeOp::Background.

This method will return an error if the image is not animated.

pub fn finish(self) -> Result<(), EncodingError>

Consume the stream writer with validation.

Unlike a simple drop this ensures that the all data was written correctly. When other validation options (chunk sequencing) had been turned on in the configuration of inner Writer, then it will also do a check on their correctness. Differently from Writer::finish, this just flushes, returns error if some data is abandoned.