Struct png::Transformations

pub struct Transformations {
    // some fields omitted
}

Output transformations

Only TRANSFORM_IDENTITY and TRANSFORM_EXPAND | TRANSFORM_STRIP_ALPHA can be used at the moment.

Methods

impl Transformations

fn empty() -> Transformations

Returns an empty set of flags.

fn all() -> Transformations

Returns the set containing all flags.

fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

fn from_bits(bits: u32) -> Option<Transformations>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u32) -> Transformations

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Transformations) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Transformations) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Transformations)

Inserts the specified flags in-place.

fn remove(&mut self, other: Transformations)

Removes the specified flags in-place.

fn toggle(&mut self, other: Transformations)

Toggles the specified flags in-place.