png
Module chunk

Chunk types and functions

Structs

ChunkType

Constants

IDAT

Image data

IEND

Image trailer

IHDR

Image header

PLTE

Palette

acTL

Animation control

bKGD

Background colour

cHRM

Source system’s pixel chromaticities

cICP

Coding-independent code points for video signal type identification chunk

cLLI

Content Light Level Information chunk

eXIf

EXIF metadata chunk

fcTL

Frame control

fdAT

Frame data

gAMA

Source system’s gamma value

iCCP

ICC profile chunk

iTXt

UTF-8 textual data

mDCV

Mastering Display Color Volume chunk

pHYs

Physical pixel dimensions

sBIT

sRGB

sRGB color space chunk

tEXt

Latin-1 uncompressed textual data

tIME

Image last-modification time

tRNS

Transparency

zTXt

Latin-1 compressed textual data

Functions

is_critical

Returns true if the chunk is critical.

is_private

Returns true if the chunk is private.

reserved_set

Checks whether the reserved bit of the chunk name is set. If it is set the chunk name is invalid.

safe_to_copy

Returns true if the chunk is safe to copy if unknown.