Module png::chunk

Chunk types and functions

Constants

IDAT

Image data

IEND

Image trailer

IHDR

Image header

PLTE

Palette

acTL

Animation control

bKGD

Background colour

fcTL

Frame control

fdAT

Frame data

pHYs

Physical pixel dimensions

tIME

Image last-modification time

tRNS

Transparency

Functions

is_critical

Returns true if the chunk is critical.

is_private

Returns true if the chunk is private.

reserved_set

Checks whether the reserved bit of the chunk name is set. If it is set the chunk name is invalid.

safe_to_copy

Returns true if the chunk is safe to copy if unknown.

Type Definitions

ChunkType