Module png::common

Common types shared between the encoder and decoder

Modules

deflate_convert
Mod to encapsulate the converters depending on the deflate crate.

Structs

AnimationControl
Animation control information

FrameControl
Frame control information

Info
PNG info struct

PixelDimensions
Pixel dimensions information

Transformations
Output transformations

Enums

BitDepth
Bit depth of the png file

BlendOp
How pixels are written into the buffer.

BytesPerPixel
Internal count of bytes per pixel. This is used for filtering which never uses sub-byte units. This essentially reduces the number of possible byte chunk lengths to a very small set of values appropriate to be defined as an enum.

ColorType
Describes the layout of samples in a pixel

Compression
The type and strength of applied compression.

DisposeOp
How to reset buffer of an animated png (APNG) at the end of a frame.

Unit
Physical unit of the pixel dimensions