Module png::common

Common types shared between the encoder and decoder

Structs

AnimationControl

Animation control information

FrameControl

Frame control information

Info

PNG info struct

ParameterError

PixelDimensions

Pixel dimensions information

ScaledFloat

An unsigned integer scaled version of a floating point value, equivalent to an integer quotient with fixed denominator (100_000)).

SourceChromaticities

Chromaticities of the color space primaries

Transformations

Output transformations

Enums

BitDepth

Bit depth of the PNG file. Specifies the number of bits per sample.

BlendOp

How pixels are written into the buffer.

BytesPerPixel 🔒

Internal count of bytes per pixel. This is used for filtering which never uses sub-byte units. This essentially reduces the number of possible byte chunk lengths to a very small set of values appropriate to be defined as an enum.

ColorType

Describes how a pixel is encoded.

Compression

The type and strength of applied compression.

DisposeOp

How to reset buffer of an animated png (APNG) at the end of a frame.

ParameterErrorKind 🔒

SrgbRenderingIntent

The rendering intent for an sRGB image.

Unit

Physical unit of the pixel dimensions