png

Module decoder

Re-exports

pub use interlace_info::InterlaceInfo;

Modules

interlace_info 🔒

read_decoder 🔒

stream 🔒

transform 🔒

Transforming a decompressed, unfiltered row into the final output.

unfiltering_buffer 🔒

zlib 🔒

Structs

Decoder

PNG Decoder

InterlacedRow

A row of data with interlace information attached.

Limits

Limits on the resources the Decoder is allowed too use

OutputInfo

Output info.

Reader

PNG reader (mostly high-level interface)

Row

A row of data without interlace information.

SubframeInfo 🔒

The subframe specific information.