Struct png::decoder::stream::StreamingDecoder

pub struct StreamingDecoder {
    state: Option<State>,
    current_chunk: (Crc32, u32, Vec<u8>),
    inflater: InflateStream,
    info: Option<Info>,
    current_seq_no: Option<u32>,
    have_idat: bool,
}

PNG StreamingDecoder (low-level interface)

Fields
state: Option<State> current_chunk: (Crc32, u32, Vec<u8>) inflater: InflateStream info: Option<Info> current_seq_no: Option<u32> have_idat: bool
Methods

impl StreamingDecoder

fn new() -> StreamingDecoder

Creates a new StreamingDecoder

Allocates the internal buffers.

fn reset(&mut self)

Resets the StreamingDecoder

fn update<'a>(&'a mut self,
buf: &[u8])
-> Result<(usize, Decoded<'a>), DecodingError>

Low level StreamingDecoder interface.

Allows to stream partial data to the encoder. Returns a tuple containing the bytes that have been consumed from the input buffer and the current decoding result.

fn next_state<'a>(&'a mut self,
buf: &[u8])
-> Result<(usize, Decoded<'a>), DecodingError>

fn parse_chunk(&mut self, type_str: [u8; 4]) -> Result<Decoded, DecodingError>

fn get_info_or_err(&self) -> Result<&Info, DecodingError>

fn parse_fctl(&mut self) -> Result<Decoded, DecodingError>

fn parse_actl(&mut self) -> Result<Decoded, DecodingError>

fn parse_plte(&mut self) -> Result<Decoded, DecodingError>

fn parse_trns(&mut self) -> Result<Decoded, DecodingError>

fn parse_phys(&mut self) -> Result<Decoded, DecodingError>

fn parse_ihdr(&mut self) -> Result<Decoded, DecodingError>