Module png::decoder::transform

Transforming a decompressed, unfiltered row into the final output.

Modules

palette 🔒

Helpers for taking a slice of indices (indices into PLTE and/or trNS entries) and transforming this into RGB or RGBA output.

Functions

copy_row 🔒

create_transform_fn

Returns a transformation function that should be applied to image rows based on 1) decoded image metadata (info) and 2) the transformations requested by the crate client (transform).

expand_gray_u8 🔒

expand_gray_u8_with_trns 🔒

expand_trns_and_strip_line16 🔒

expand_trns_line 🔒

expand_trns_line16 🔒

transform_row_strip16 🔒

unpack_bits 🔒

Type Aliases

TransformFn

Type of a function that can transform a decompressed, unfiltered row (the 1st argument) into the final pixels (the 2nd argument), optionally using image metadata (e.g. PLTE data can be accessed using the 3rd argument).