Module png::encoder

Structs

ChunkWriter 🔒

This writer is used between the actual writer and the ZlibEncoder and has the job of packaging the compressed data into a PNG chunk, based on the image metadata

Encoder

PNG Encoder.

FormatError

Options 🔒

Decoding options, internal type, forwarded to the Writer.

PartialInfo 🔒

Contains the subset of attributes of Info needed for Writer to function

StreamWriter

Streaming PNG writer

Writer

PNG writer

Enums

ChunkOutput 🔒

EncodingError

FormatErrorKind 🔒

Wrapper 🔒

This enum is used to be allow the StreamWriter to keep its inner ChunkWriter without wrapping it inside a ZlibEncoder. This is used in the case that between the change of state that happens when the last write of a frame is performed an error occurs, which obviously has to be returned. This creates the problem of where to store the writer before exiting the function, and this is where Wrapper comes in.

Constants

DEFAULT_BUFFER_LENGTH 🔒

Functions

write_chunk 🔒

Type Aliases

Result