Struct png::encoder::PartialInfo

struct PartialInfo {
    width: u32,
    height: u32,
    bit_depth: BitDepth,
    color_type: ColorType,
    frame_control: Option<FrameControl>,
    animation_control: Option<AnimationControl>,
    compression: Compression,
    has_palette: bool,
}

Contains the subset of attributes of Info needed for Writer to function

Fields
width: u32height: u32bit_depth: BitDepthcolor_type: ColorTypeframe_control: Option<FrameControl>animation_control: Option<AnimationControl>compression: Compressionhas_palette: bool
Implementations

impl PartialInfo

fn new(info: &Info<'_>) -> Self

fn bpp_in_prediction(&self) -> BytesPerPixel

fn raw_row_length(&self) -> usize

fn raw_row_length_from_width(&self, width: u32) -> usize

fn to_info(&self) -> Info<'static>

Converts this partial info to an owned Info struct, setting missing values to their defaults