Struct png::FrameControl

pub struct FrameControl {
    pub sequence_number: u32,
    pub width: u32,
    pub height: u32,
    pub x_offset: u32,
    pub y_offset: u32,
    pub delay_num: u16,
    pub delay_den: u16,
    pub dispose_op: u8,
    pub blend_op: u8,
}

Frame control information

Fields
sequence_number: u32

Sequence number of the animation chunk, starting from 0

width: u32

Width of the following frame

height: u32

Height of the following frame

x_offset: u32

X position at which to render the following frame

y_offset: u32

Y position at which to render the following frame

delay_num: u16

Frame delay fraction numerator

delay_den: u16

Frame delay fraction denominator

dispose_op: u8

Type of frame area disposal to be done after rendering this frame

blend_op: u8

Type of frame area rendering for this frame