Struct png::Info

#[non_exhaustive]
pub struct Info<'a> {

    pub width: u32,
    pub height: u32,
    pub bit_depth: BitDepth,
    pub color_type: ColorType,
    pub interlaced: bool,
    pub trns: Option<Cow<'a, [u8]>>,
    pub pixel_dims: Option<PixelDimensions>,
    pub palette: Option<Cow<'a, [u8]>>,
    pub gama_chunk: Option<ScaledFloat>,
    pub chrm_chunk: Option<SourceChromaticities>,
    pub frame_control: Option<FrameControl>,
    pub animation_control: Option<AnimationControl>,
    pub compression: Compression,
    pub source_gamma: Option<ScaledFloat>,
    pub source_chromaticities: Option<SourceChromaticities>,
    pub srgb: Option<SrgbRenderingIntent>,
    pub icc_profile: Option<Cow<'a, [u8]>>,
    pub uncompressed_latin1_text: Vec<TEXtChunk>,
    pub compressed_latin1_text: Vec<ZTXtChunk>,
    pub utf8_text: Vec<ITXtChunk>,

}

PNG info struct

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

width: u32height: u32bit_depth: BitDepthcolor_type: ColorType

How colors are stored in the image.

interlaced: booltrns: Option<Cow<'a, [u8]>>

The imageโs tRNS chunk, if present; contains the alpha channel of the imageโs palette, 1 byte per entry.

pixel_dims: Option<PixelDimensions>palette: Option<Cow<'a, [u8]>>

The imageโs PLTE chunk, if present; contains the RGB channels (in that order) of the imageโs palettes, 3 bytes per entry (1 per channel).

gama_chunk: Option<ScaledFloat>

The contents of the imageโs gAMA chunk, if present. Prefer source_gamma to also get the derived replacement gamma from sRGB chunks.

chrm_chunk: Option<SourceChromaticities>

The contents of the imageโs cHRM chunk, if present. Prefer source_chromaticities to also get the derived replacements from sRGB chunks.

frame_control: Option<FrameControl>animation_control: Option<AnimationControl>compression: Compressionsource_gamma: Option<ScaledFloat>

Gamma of the source system. Set by both gAMA as well as to a replacement by sRGB chunk.

source_chromaticities: Option<SourceChromaticities>

Chromaticities of the source system. Set by both cHRM as well as to a replacement by sRGB chunk.

srgb: Option<SrgbRenderingIntent>

The rendering intent of an SRGB image.

Presence of this value also indicates that the image conforms to the SRGB color space.

icc_profile: Option<Cow<'a, [u8]>>

The ICC profile for the image.

uncompressed_latin1_text: Vec<TEXtChunk>

tEXt field

compressed_latin1_text: Vec<ZTXtChunk>

zTXt field

utf8_text: Vec<ITXtChunk>

iTXt field

Implementations

impl Info<'_>

pub fn with_size(width: u32, height: u32) -> Self

A utility constructor for a default info with width and height.

pub fn size(&self) -> (u32, u32)

Size of the image, width then height.

pub fn is_animated(&self) -> bool

Returns true if the image is an APNG image.

pub fn animation_control(&self) -> Option<&AnimationControl>

Returns the frame control information of the image.

pub fn frame_control(&self) -> Option<&FrameControl>

Returns the frame control information of the current frame

pub fn bits_per_pixel(&self) -> usize

Returns the number of bits per pixel.

pub fn bytes_per_pixel(&self) -> usize

Returns the number of bytes per pixel.

pub fn raw_bytes(&self) -> usize

Returns the number of bytes needed for one deinterlaced image.

pub fn raw_row_length(&self) -> usize

Returns the number of bytes needed for one deinterlaced row.

pub fn raw_row_length_from_width(&self, width: u32) -> usize

Returns the number of bytes needed for one deinterlaced row of width width.

pub fn encode<W: Write>(&self, w: W) -> Result<(), EncodingError>

Encode this header to the writer.

Note that this does not include the PNG signature, it starts with the IHDR chunk and then includes other chunks that were added to the header.