Struct png::Reader

pub struct Reader<R: Read> { /* fields omitted */ }

PNG reader (mostly high-level interface)

Provides a high level that iterates over lines or whole images.

Methods

impl<R: Read> Reader<R>

pub fn info(&self) -> &Info

pub fn next_frame(&mut self, buf: &mut [u8]) -> Result<(), DecodingError>

Decodes the next frame into buf

pub fn next_row(&mut self) -> Result<Option<&[u8]>, DecodingError>

Returns the next processed row of the image

pub fn next_interlaced_row(
&mut self
) -> Result<Option<(&[u8], Option<(u8, u32, u32)>)>, DecodingError>

Returns the next processed row of the image

pub fn output_color_type(&mut self) -> (ColorType, BitDepth)

Returns the color type and the number of bits per sample of the data returned by Reader::next_row and Reader::frames`.

pub fn output_buffer_size(&self) -> usize

Returns the number of bytes required to hold a deinterlaced image frame that is decoded using the given input transformations.

pub fn output_line_size(&self, width: u32) -> usize

Returns the number of bytes required to hold a deinterlaced row.