Struct png::StreamingDecoder

pub struct StreamingDecoder { /* fields omitted */ }

PNG StreamingDecoder (low-level interface)

Methods

impl StreamingDecoder

pub fn new() -> StreamingDecoder

Creates a new StreamingDecoder

Allocates the internal buffers.

pub fn reset(&mut self)

Resets the StreamingDecoder

pub fn update<'a>(
&'a mut self,
buf: &[u8]
) -> Result<(usize, Decoded<'a>), DecodingError>

Low level StreamingDecoder interface.

Allows to stream partial data to the encoder. Returns a tuple containing the bytes that have been consumed from the input buffer and the current decoding result.